update review, move credssp sequence to another module
